feat(yi-connect): rewire coordinator + sub-chapter pages to Supabase auth
Folds /coordinator/* and /chapter-lead/* pages into the (dashboard) route
group. Replaces legacy cookie-based session lookups with requireRole() +
user.id -> stakeholder_coordinators / sub_chapter_leads lookups.

Split client form pages (events/new, members/add) into server shells +
client form components so sub_chapter_id is resolved server-side and
passed as a prop instead of read from document.cookie.
